+++ pkg/apis/deployment/v2alpha1/license_spec.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2025 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2026 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -21,6 +21,8 @@
 package v2alpha1
 
 import (
+	"time"
+
 	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	shared "github.com/arangodb/kube-arangodb/pkg/apis/shared"
@@ -29,7 +31,9 @@ import (
 )
 
 const (
-	LicenseExpirationGraceRatio = 0.75
+	LicenseExpirationMaxGraceRatio  = 0.3333
+	LicenseExpirationMinGraceRatio  = 0.6667
+	LicenseExpirationMinGracePeriod = 15 * time.Minute
 )
 
 type LicenseMode string