+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - (Feature) (Platform) Package Copy CLI
 - (Feature) (Platform) MetaV1 GetBatch
 - (Feature) Switch Endpoints to EndpointSlices
+- (Feature) Ensure License Grace Period
 
 ## [1.4.1](https://github.com/arangodb/kube-arangodb/tree/1.4.1) (2026-01-27)
 - (Bugfix) (Platform) Ensure Helm History limit in ArangoPlatformService